Moved %s and %p functions to separate files.

+++ src/ft_putnbr_hex_fd.c
@@ -6,15 +6,15 @@
 /*   By: ccamargo <ccamargo@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/10 14:39:35 by ccamargo          #+#    #+#             */
-/*   Updated: 2022/06/12 17:32:49 by ccamargo         ###   ########.fr       */
+/*   Updated: 2022/06/12 20:44:51 by ccamargo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_putnbr_hex_fd(unsigned long nbr, char *base, int fd)
+int	ft_putnbr_hex_fd(unsigned long nbr, char *base, int fd)
 {
-	size_t	count;
+	int	count;
 
 	count = 0;
 	if (nbr / 16 > 0)
